Rename cssStatementStringBeforeBlock to beforeBlockString
to align with other `*String` utils
