+++ src/rules/block-opening-brace-newline-after/index.js
@@ -2,7 +2,7 @@ import {
   blockString,
   hasBlock,
   hasEmptyBlock,
-  cssStatementStringBeforeBlock,
+  beforeBlockString,
   rawNodeString,
   report,
   ruleMessages,
@@ -61,7 +61,7 @@ export default function (expectation) {
           report({
             message: m,
             node: statement,
-            index: cssStatementStringBeforeBlock(statement, { noBefore: true }).length + 1,
+            index: beforeBlockString(statement, { noBefore: true }).length + 1,
             result,
             ruleName,
           })
