+++ src/rules/number-no-trailing-zeros/index.js
@@ -3,7 +3,7 @@ import {
   blurComments,
   blurFunctionArguments,
   hasBlock,
-  cssStatementStringBeforeBlock,
+  beforeBlockString,
   report,
   ruleMessages,
   validateOptions,
@@ -30,7 +30,7 @@ export default function (actual) {
       if (atRule.name === "import") { return }
 
       const source = (hasBlock(atRule))
-        ? cssStatementStringBeforeBlock(atRule, { noBefore: true })
+        ? beforeBlockString(atRule, { noBefore: true })
         : atRule.toString()
       check(source, atRule)
     })
