+++ src/utils/__tests__/beforeBlockString-test.js
@@ -1,15 +1,15 @@
 import test from "tape"
 import postcss from "postcss"
-import cssStatementStringBeforeBlock from "../cssStatementStringBeforeBlock"
+import beforeBlockString from "../beforeBlockString"
 
-test("cssStatementStringBeforeBlock rules", t => {
+test("beforeBlockString rules", t => {
   t.equal(postcssCheck("a {}"), "a ")
   t.equal(postcssCheck("\na\n{}"), "\na\n")
   t.equal(postcssCheck("\n\na,\nb,\n\tspan > .foo\n{}"), "\n\na,\nb,\n\tspan > .foo\n")
   t.end()
 })
 
-test("cssStatementStringBeforeBlock at-rules", t => {
+test("beforeBlockString at-rules", t => {
   t.equal(postcssCheck("@media print {}"), "@media print ")
   t.equal(postcssCheck("\n@media print, screen\n\t{}"), "\n@media print, screen\n\t")
   t.equal(postcssCheck("@supports (animation-name: test) {}"), "@supports (animation-name: test) ")
@@ -25,13 +25,13 @@ test("cssStatementStringBeforeBlock at-rules", t => {
   t.end()
 })
 
-test("cssStatementStringBeforeBlock with noBefore", t => {
+test("beforeBlockString with noBefore", t => {
   t.equal(postcssCheck({ noBefore: true }, "\na {}"), "a ")
   t.equal(postcssCheck({ noBefore: true }, "\n@media print {}"), "@media print ")
   t.end()
 })
 
-test("cssStatementStringBeforeBlock with declaration directly at root", t => {
+test("beforeBlockString with declaration directly at root", t => {
   t.equal(postcssCheck("foo: bar;"), "")
   t.end()
 })
@@ -41,5 +41,5 @@ function postcssCheck(options={}, cssString) {
     cssString = options
   }
   const root = postcss.parse(cssString)
-  return cssStatementStringBeforeBlock(root.first, options)
+  return beforeBlockString(root.first, options)
 }
