+++ src/utils/index.js
@@ -6,7 +6,7 @@ export { default as cssNodeContextLookup } from "./cssNodeContextLookup"
 export { default as blockString } from "./blockString"
 export { default as hasBlock } from "./hasBlock"
 export { default as hasEmptyBlock } from "./hasEmptyBlock"
-export { default as cssStatementStringBeforeBlock } from "./cssStatementStringBeforeBlock"
+export { default as beforeBlockString } from "./beforeBlockString"
 export { default as cssWordIsVariable } from "./cssWordIsVariable"
 export { default as declarationValueIndexOffset } from "./declarationValueIndexOffset"
 export { default as findAtRuleContext } from "./findAtRuleContext"
